Update token id display
